consume a bunch of notifications at once
Sometimes I'm seeing in logs that writes to sockets that are used for
kicking threads return EWOULDBLOCK. This patch helps this case by
consuming more than one notification at a time.

memcached_shutdown check is moved after read so that shutdown notification
is not lost.

Also thread_libevent_process tries to process number of incoming
connections that's equal to number of notifications consumed. This is
necessary to avoid losing new connections.